+++ packages/components/CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `Divider`, `Button Group`, `Carousel` and `Text List` match new design language
 - Background color of `Scroller`'s `prev` and `next` buttons in dark theme
 - Partials: Removed deprecated `withoutTags` option for all partials, please use `format: 'jsx'` instead
+- `Content Wrapper` default value of prop `width` has changed from `basic` to `extended`
 
 #### Added